collect index formatting tests
