+++ pandas/tests/indexes/period/test_ops.py
@@ -115,164 +115,6 @@ class TestPeriodIndexOps(Ops):
             tm.assert_raises_regex(
                 ValueError, errmsg, np.argmax, pr, out=0)
 
-    def test_representation(self):
-        # GH 7601
-        idx1 = PeriodIndex([], freq='D')
-        idx2 = PeriodIndex(['2011-01-01'], freq='D')
-        idx3 = PeriodIndex(['2011-01-01', '2011-01-02'], freq='D')
-        idx4 = PeriodIndex(['2011-01-01', '2011-01-02', '2011-01-03'],
-                           freq='D')
-        idx5 = PeriodIndex(['2011', '2012', '2013'], freq='A')
-        idx6 = PeriodIndex(['2011-01-01 09:00', '2012-02-01 10:00',
-                            'NaT'], freq='H')
-        idx7 = pd.period_range('2013Q1', periods=1, freq="Q")
-        idx8 = pd.period_range('2013Q1', periods=2, freq="Q")
-        idx9 = pd.period_range('2013Q1', periods=3, freq="Q")
-        idx10 = PeriodIndex(['2011-01-01', '2011-02-01'], freq='3D')
-
-        exp1 = """PeriodIndex([], dtype='period[D]', freq='D')"""
-
-        exp2 = """PeriodIndex(['2011-01-01'], dtype='period[D]', freq='D')"""
-
-        exp3 = ("PeriodIndex(['2011-01-01', '2011-01-02'], dtype='period[D]', "
-                "freq='D')")
-
-        exp4 = ("PeriodIndex(['2011-01-01', '2011-01-02', '2011-01-03'], "
-                "dtype='period[D]', freq='D')")
-
-        exp5 = ("PeriodIndex(['2011', '2012', '2013'], dtype='period[A-DEC]', "
-                "freq='A-DEC')")
-
-        exp6 = ("PeriodIndex(['2011-01-01 09:00', '2012-02-01 10:00', 'NaT'], "
-                "dtype='period[H]', freq='H')")
-
-        exp7 = ("PeriodIndex(['2013Q1'], dtype='period[Q-DEC]', "
-                "freq='Q-DEC')")
-
-        exp8 = ("PeriodIndex(['2013Q1', '2013Q2'], dtype='period[Q-DEC]', "
-                "freq='Q-DEC')")
-
-        exp9 = ("PeriodIndex(['2013Q1', '2013Q2', '2013Q3'], "
-                "dtype='period[Q-DEC]', freq='Q-DEC')")
-
-        exp10 = ("PeriodIndex(['2011-01-01', '2011-02-01'], "
-                 "dtype='period[3D]', freq='3D')")
-
-        for idx, expected in zip([idx1, idx2, idx3, idx4, idx5,
-                                  idx6, idx7, idx8, idx9, idx10],
-                                 [exp1, exp2, exp3, exp4, exp5,
-                                  exp6, exp7, exp8, exp9, exp10]):
-            for func in ['__repr__', '__unicode__', '__str__']:
-                result = getattr(idx, func)()
-                assert result == expected
-
-    def test_representation_to_series(self):
-        # GH 10971
-        idx1 = PeriodIndex([], freq='D')
-        idx2 = PeriodIndex(['2011-01-01'], freq='D')
-        idx3 = PeriodIndex(['2011-01-01', '2011-01-02'], freq='D')
-        idx4 = PeriodIndex(['2011-01-01', '2011-01-02',
-                            '2011-01-03'], freq='D')
-        idx5 = PeriodIndex(['2011', '2012', '2013'], freq='A')
-        idx6 = PeriodIndex(['2011-01-01 09:00', '2012-02-01 10:00',
-                            'NaT'], freq='H')
-
-        idx7 = pd.period_range('2013Q1', periods=1, freq="Q")
-        idx8 = pd.period_range('2013Q1', periods=2, freq="Q")
-        idx9 = pd.period_range('2013Q1', periods=3, freq="Q")
-
-        exp1 = """Series([], dtype: object)"""
-
-        exp2 = """0   2011-01-01
-dtype: object"""
-
-        exp3 = """0   2011-01-01
-1   2011-01-02
-dtype: object"""
-
-        exp4 = """0   2011-01-01
-1   2011-01-02
-2   2011-01-03
-dtype: object"""
-
-        exp5 = """0   2011
-1   2012
-2   2013
-dtype: object"""
-
-        exp6 = """0   2011-01-01 09:00
-1   2012-02-01 10:00
-2                NaT
-dtype: object"""
-
-        exp7 = """0   2013Q1
-dtype: object"""
-
-        exp8 = """0   2013Q1
-1   2013Q2
-dtype: object"""
-
-        exp9 = """0   2013Q1
-1   2013Q2
-2   2013Q3
-dtype: object"""
-
-        for idx, expected in zip([idx1, idx2, idx3, idx4, idx5,
-                                  idx6, idx7, idx8, idx9],
-                                 [exp1, exp2, exp3, exp4, exp5,
-                                  exp6, exp7, exp8, exp9]):
-            result = repr(pd.Series(idx))
-            assert result == expected
-
-    def test_summary(self):
-        # GH9116
-        idx1 = PeriodIndex([], freq='D')
-        idx2 = PeriodIndex(['2011-01-01'], freq='D')
-        idx3 = PeriodIndex(['2011-01-01', '2011-01-02'], freq='D')
-        idx4 = PeriodIndex(
-            ['2011-01-01', '2011-01-02', '2011-01-03'], freq='D')
-        idx5 = PeriodIndex(['2011', '2012', '2013'], freq='A')
-        idx6 = PeriodIndex(
-            ['2011-01-01 09:00', '2012-02-01 10:00', 'NaT'], freq='H')
-
-        idx7 = pd.period_range('2013Q1', periods=1, freq="Q")
-        idx8 = pd.period_range('2013Q1', periods=2, freq="Q")
-        idx9 = pd.period_range('2013Q1', periods=3, freq="Q")
-
-        exp1 = """PeriodIndex: 0 entries
-Freq: D"""
-
-        exp2 = """PeriodIndex: 1 entries, 2011-01-01 to 2011-01-01
-Freq: D"""
-
-        exp3 = """PeriodIndex: 2 entries, 2011-01-01 to 2011-01-02
-Freq: D"""
-
-        exp4 = """PeriodIndex: 3 entries, 2011-01-01 to 2011-01-03
-Freq: D"""
-
-        exp5 = """PeriodIndex: 3 entries, 2011 to 2013
-Freq: A-DEC"""
-
-        exp6 = """PeriodIndex: 3 entries, 2011-01-01 09:00 to NaT
-Freq: H"""
-
-        exp7 = """PeriodIndex: 1 entries, 2013Q1 to 2013Q1
-Freq: Q-DEC"""
-
-        exp8 = """PeriodIndex: 2 entries, 2013Q1 to 2013Q2
-Freq: Q-DEC"""
-
-        exp9 = """PeriodIndex: 3 entries, 2013Q1 to 2013Q3
-Freq: Q-DEC"""
-
-        for idx, expected in zip([idx1, idx2, idx3, idx4, idx5,
-                                  idx6, idx7, idx8, idx9],
-                                 [exp1, exp2, exp3, exp4, exp5,
-                                  exp6, exp7, exp8, exp9]):
-            result = idx.summary()
-            assert result == expected
-
     def test_resolution(self):
         for freq, expected in zip(['A', 'Q', 'M', 'D', 'H',
                                    'T', 'S', 'L', 'U'],
